Crate bxcan

Driver for the STM32 bxCAN peripheral.

This crate provides a reusable driver for the bxCAN peripheral found in many low- to middle-end STM32 microcontrollers. HALs for compatible chips can reexport this crate and implement its traits to easily expose a featureful CAN driver.

Caveats:

Only RX FIFO 0 is supported, FIFO 1 will not be used.

Modules

filter
Filter bank API.

Structs

Can
Interface to a CAN peripheral.

CanConfig
Configuration proxy to be used with Can::configure().

Data
Payload of a CAN data frame.

ExtendedId
Extended 29-bit CAN Identifier (0..=1FFF_FFFF).

Frame
A CAN data or remote frame.

FramePriority
Priority of a CAN frame.

Interrupts
A set of bxCAN interrupts.

RegisterBlock
Register block of bxCAN peripherals.

Rx
Interface to the CAN receiver part.

StandardId
Standard 11-bit CAN Identifier (0..=0x7FF).

Tx
Interface to the CAN transmitter part.

Enums

Id
A CAN Identifier (standard or extended).

Interrupt
bxCAN interrupt sources.

Traits

FilterOwner
A bxCAN instance that owns filter banks.

Instance
A bxCAN peripheral instance.

MasterInstance
A bxCAN master instance that shares filter banks with a slave instance.