Struct bxcan::Can

pub struct Can<I: Instance> { /* fields omitted */ }

Interface to a CAN peripheral.

Implementations

impl<I> Can<I> where
I: Instance,

pub fn new(instance: I) -> Self

Creates a CAN interface, taking ownership of the raw peripheral.

pub fn instance(&mut self) -> &mut I

Returns a reference to the peripheral instance.

This allows accessing HAL-specific data stored in the instance type.

pub fn configure<F>(&mut self, f: F) where
F: FnOnce(&mut CanConfig<'_, I>),

Configure bit timings and silent/loop-back mode.

Acutal configuration happens on the CanConfig that is passed to the closure. It must be done this way because those configuration bits can only be set if the CAN controller is in a special init mode. Puts the peripheral in sleep mode afterwards. Can::enable() must be called to exit sleep mode and start reception and transmission.

pub fn modify_config(&mut self) -> CanConfig<'_, I>

Configure bit timings and silent/loop-back mode.

pub fn set_automatic_wakeup(&mut self, enabled: bool)

Configures the automatic wake-up feature.

pub fn enable(&mut self) -> Result<(), Infallible>

Start reception and transmission.

Waits for 11 consecutive recessive bits to sync to the CAN bus.

pub fn sleep(&mut self)

Puts the peripheral in a sleep mode to save power.

Reception and transmission is disabled.

pub fn enable_interrupt(&mut self, interrupt: Interrupt)

Starts listening for a CAN interrupt.

pub fn enable_interrupts(&mut self, interrupts: Interrupts)

Starts listening for a set of CAN interrupts.

pub fn disable_interrupt(&mut self, interrupt: Interrupt)

Stops listening for a CAN interrupt.

pub fn disable_interrupts(&mut self, interrupts: Interrupts)

Stops listening for a set of CAN interrupts.

pub fn clear_wakeup_interrupt(&mut self)

Clears the pending flag of Interrupt::Wakeup.

pub fn transmit(&mut self, frame: &Frame) -> Result<Option<Frame>, Infallible>

Puts a CAN frame in a free transmit mailbox for transmission on the bus.

Frames are transmitted to the bus based on their priority (identifier). Transmit order is preserved for frames with identical identifiers. If all transmit mailboxes are full, a higher priority frame replaces the lowest priority frame, which is returned as Ok(Some(frame)).

pub fn is_transmitter_idle(&self) -> bool

Returns true if no frame is pending for transmission.

pub fn receive(&mut self) -> Result<Frame, ()>

Returns a received frame if available.

Returns Err when a frame was lost due to buffer overrun.

pub fn split_by_ref(&mut self) -> (&mut Tx<I>, &mut Rx<I>)

Splits this Can instance into transmitting and receiving halves, by reference.

pub fn split(self) -> (Tx<I>, Rx<I>)

Consumes this Can instance and splits it into transmitting and receiving halves.

impl<I: FilterOwner> Can<I>

pub fn modify_filters(&mut self) -> MasterFilters<'_, I>

Accesses the filter banks owned by this CAN peripheral.

To modify filters of a slave peripheral, modify_filters has to be called on the master peripheral instead.