Struct bxcan::Frame

pub struct Frame { /* fields omitted */ }

A CAN data or remote frame.

Implementations

impl Frame

pub fn new_data(id: impl Into<Id>, data: impl Into<Data>) -> Self

Creates a new data frame.

pub fn new_remote(id: impl Into<Id>, dlc: u8) -> Self

Creates a new remote frame with configurable data length code (DLC).

Panics

This function will panic if dlc is not inside the valid range 0..=8.

pub fn is_extended(&self) -> bool

Returns true if this frame is an extended frame.

pub fn is_standard(&self) -> bool

Returns true if this frame is a standard frame.

pub fn is_remote_frame(&self) -> bool

Returns true if this frame is a remote frame.

pub fn is_data_frame(&self) -> bool

Returns true if this frame is a data frame.

pub fn id(&self) -> Id

Returns the frame identifier.

pub fn priority(&self) -> FramePriority

Returns the priority of this frame.

pub fn dlc(&self) -> u8

Returns the data length code (DLC) which is in the range 0..8.

For data frames the DLC value always matches the length of the data. Remote frames do not carry any data, yet the DLC can be greater than 0.

pub fn data(&self) -> Option<&Data>

Returns the frame data (0..8 bytes in length) if this is a data frame.

If this is a remote frame, returns None.